Changed the overall GP time variable to a double from a float Added the overall GP time to the GP end screen
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1452 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/race_manager.hpp
@@ -40,7 +40,7 @@ public:
     virtual std::string getKartName(int kart) const { return "";}
     virtual int  getPositionScore(int pos ) const {return 0;}
     virtual void addKartOverallTime (int kart, float race_time) {}
-    virtual float getKartOverallTime (int kart) const {return 0.0f;}
+    virtual double getKartOverallTime (int kart) const {return 0.0f;}
 };
 
 class GrandPrixMode : public RaceMode
@@ -55,13 +55,13 @@ private:
     {
         std::string ident;//The .tkkf filename without the .tkkf
         int score;
-        float overall_time;
+        double overall_time;
         int prev_finish_pos;
         int player;//Which player controls the kart, for the AI this is
         //the number of players.
 
         KartStatus(const std::string& ident_, const int& score_,
-                   const float& overall_time_, const int& prev_finish_pos_, 
+                   const double& overall_time_, const int& prev_finish_pos_, 
                    const int& player_) :
                 ident(ident_), score(score_), overall_time(overall_time_),
                  prev_finish_pos(prev_finish_pos_), player(player_){}
@@ -92,7 +92,7 @@ public:
         m_karts[kart].score +=
             getPositionScore(pos);
     }
-    float getKartOverallTime (int kart) const {return m_karts[kart].overall_time;}
+    double getKartOverallTime (int kart) const {return m_karts[kart].overall_time;}
     void addKartOverallTime  (int kart, float race_time)
     {
 		m_karts[kart].overall_time += race_time;
@@ -182,7 +182,7 @@ public:
     std::string getKartName(int kart) const { return m_mode->getKartName(kart);  }
     void addKartScore(int kart, int pos)  { m_mode->addKartScore(kart, pos);     }
     void addKartOverallTime(int kart, float race_time) {m_mode->addKartOverallTime(kart, race_time); }
-    float getKartOverallTime(int kart) {return m_mode->getKartOverallTime(kart); }
+    double getKartOverallTime(int kart) {return m_mode->getKartOverallTime(kart); }
     void addFinishedKarts(int num)        { m_num_finished_karts += num;         }
     void PlayerFinishes()                 { m_num_finished_players++;            }
     int  allPlayerFinished()    {return m_num_finished_players==m_players.size();}
